refactor: 리뷰 반영
- renderRouteSystem -> useRouteSystem 변경
- useNavigate 반영
