+++ packages/index.ts
@@ -10,7 +10,7 @@ import {
   XhrUploadStateArgs,
 } from './http-api';
 import { cache } from './proxies';
-import { ModuleRouteModel, RenderRouteSystem } from './route-system';
+import { ModuleRouteModel, useRouteSystem } from './route-system';
 import validate, {
   mergeValidates,
   ValidateResultModel,
@@ -37,7 +37,7 @@ export {
   headerPipe,
   createHttpHeaderProvider,
   cache,
-  RenderRouteSystem,
+  useRouteSystem,
   validate,
   mergeValidates,
   ValidateResultModel,
