+++ packages/route-system/index.ts
@@ -1,2 +1,2 @@
-export * from './renderRouteSystem';
-export * from './renderRouteSystem.model';
+export * from './useRouteSystem';
+export * from './useRouteSystem.model';
